requirements.txt
requests==2.31.0
pytest==7.4.3
pytest-html==4.1.1
allure-pytest==2.13.2
openpyxl==3.1.2
PyYAML==6.0.1

